refactored code
code refactoring, made simple blocks for understanding and classes to simplification
